LPD-35857 portal-search-web-test: Move test to the right modulet
https://liferay.atlassian.net/browse/LPD-35857
