+++ modules/apps/portal-search/portal-search-web-test/build.gradle
@@ -1,5 +1,8 @@
 dependencies {
+	testIntegrationImplementation project(":apps:asset:asset-test-util")
 	testIntegrationImplementation project(":apps:export-import:export-import-api")
+	testIntegrationImplementation project(":apps:export-import:export-import-test-util")
+	testIntegrationImplementation project(":apps:layout:layout-test-util")
 	testIntegrationImplementation project(":apps:portal-search:portal-search-test-util")
 	testIntegrationImplementation project(":apps:portal-search:portal-search-web")
 	testIntegrationImplementation project(":apps:portal-search:portal-search-web-api")
